Nock for fetch-mock for PR already assigned case
